feat(api): refactor
