+++ apps/api/src/app/workflows-v2/util/template-parser/parser-types.ts
@@ -1,0 +1,10 @@
+export type InvalidVariable = {
+  context: string;
+  message: string;
+  variable: string;
+};
+
+export type TemplateParseResult = {
+  validVariables: string[];
+  invalidVariables: InvalidVariable[];
+};
